% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanhtml.R
\name{cleanhtml}
\alias{cleanhtml}
\title{Extracts the location data from the html string}
\usage{
cleanhtml(htmlstring)
}
\arguments{
\item{htmlstring}{character vector produced by the gethtml function.  The html will be scraped from the LabCorp website.}
}
\value{
dataframe
}
\description{
This function takes a raw html string from the gethtml function and extracts the location data.  Each patient lab will include the address, city, state, zip code and lat-long coordinates for their location.
}
