% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatdata.R
\name{formatdata}
\alias{formatdata}
\title{Formats raw html into matrix}
\usage{
formatdata(df)
}
\arguments{
\item{df}{datafram producted by the \code{\link{cleanhtml}} function}
}
\value{
dataframe
}
\description{
Format a rough dataframe with location data into final form. Removes names and punctuation to leave only location values.
}
