% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gethtml.R
\name{gethtml}
\alias{gethtml}
\title{Get raw html}
\usage{
gethtml(myURL)
}
\arguments{
\item{myURL}{url string created by \code{\link{makeurl}} function}
}
\value{
character string
}
\description{
Retrieves the raw html response after passing the query string to the server
}
