% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabCorp.R
\docType{package}
\name{LabCorp}
\alias{LabCorp}
\alias{LabCorp-package}
\title{LabCorp location data}
\description{
The LabCorp package documents the process and tools used to scrape the LabCorp
website to retieve patient lab location data.  Please read the vignette for
a complete step-by-step instruction on how to use the package.
}
\examples{
vignette("LabCorp")

}
