% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabLocations.R
\docType{data}
\name{LabLocations}
\alias{LabLocations}
\title{Dataframe of service locations}
\format{dataframe}
\description{
This was created by scraping the LabCorp website. The raw html returned from the website was parsed and the location data extracted.
}
