% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querystring.R
\name{querystring}
\alias{querystring}
\title{Create a query string}
\usage{
querystring(x)
}
\arguments{
\item{x}{List with search parameters.
\itemize{
\item streetAddress
\item city
\item state - 2 character abbreviation
\item zipCode - string of 5 numeric characters
\item radius
\item testType
}}
}
\value{
string
}
\description{
This query string will become part of the search URL passed to the LabCorp servers.
}

